requirements.txt
mysql-connector-python
SQLAlchemy==2.0.3
sqlalchemy-utils
fastapi[all]~=0.89.1
fastapi-utils
PYMYSQL
uvicorn~=0.20.0
pydantic==1.10.5
cryptography
PyJWT==2.6.0
python-decouple==3.7
bcrypt

